Levy

+++ backend/app/controllers/home_details_controller.rb
@@ -1,6 +1,7 @@
 class HomeDetailsController < ApplicationController
    
-    # before_action :verify_auth, only: [:create, :update, :destroy]
+      before_action :current_user, only: [:create, :update, :destroy]
+
   # before_action :set_user, only: [:create, :update, :destroy]
 
 
@@ -10,20 +11,22 @@ class HomeDetailsController < ApplicationController
     end
 
     def create
-      home = HomeDetail.create(home_params)
-    
-      user = User.find_by(id: session[:user_id])
-      
-       if user && user.admin?
-        if home.save 
+       
+        if current_user.admin?
+        home = HomeDetail.create(home_params)
+        if home.save
+
           app_response(message: "created successfully", status: :created, data: home)
         else
           app_response(message: "failed to create", status: :unprocessable_entity, data: home.errors)
         end
-       else
-         app_response(message: "You are not authorized to perform this action", status: :unauthorized)
-       end
+      else
+        app_response(message: "You dont ave acces rihts to perform this action", status: :unauthorized)
+      end
+
     end
+    
+    
     
     
 
@@ -45,30 +48,36 @@ class HomeDetailsController < ApplicationController
   end
       
     def update
-        home = HomeDetail.find(params[:id])
-      
+      if current_user.admin?
+        home = HomeDetail.find(params[:id])      
         if home.update(home_params)
           app_response(message: "updated successfully", data: home)
         else
           app_response(message: "failed to update", status: :unprocessable_entity)
         end
+      else
+        app_response(message: "You dont ave acces rihts to perform this action", status: :unauthorized)
+      end
     end
      
     def destroy
-
+         
+      if current_user.admin?
         home = HomeDetail.find(params[:id])
-
         if home.destroy
             app_response(message: "deleted successfully")
         else
             app_response(message: "failed to delete")
         end
+      else
+        app_response(message: "You dont have acces rights to perform this action", status: :unauthorized)
+      end
     end
      
      private 
  
      def home_params
-       params.permit(:name, :description, :image_url, :address, :phone_number, :email, :location)
+       params.require(:home_detail).permit(:name, :description, :image_url, :address, :phone_number, :email, :location)
     end
